pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "saop"
version = "0.1.0"
description = "Monorepo managed by Poetry. Tooling: pre-commit, Black, Ruff, mypy."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "prometheus-client (>=0.22.1,<0.23.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "langchain-core (>=0.3.74,<0.4.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "starlette (>=0.47.2,<0.48.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "opentelemetry-api (>=1.36.0,<2.0.0)",
    "opentelemetry-sdk (>=1.36.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.36.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-requests (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-httpx (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-logging (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-grpc (>=0.57b0,<0.58)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "grpcio-tools (>=1.74.0,<2.0.0)",
    "typing-extensions (>=4.14.1,<5.0.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
    "litellm (>=1.75.8,<2.0.0)",
    "orjson (>=3.11.2,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sse-starlette (>=3.0.2,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "authlib (>=1.6.1,<2.0.0)",
    "casbin (>=1.43.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "grpcio (>=1.74.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "fastmcp (>=2.11.3,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "a2a-sdk[grpc,http-server,telemetry] (>=0.3.1,<0.4.0)",
    "mcp[cli] (>=1.13.0,<2.0.0)",
]

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.3.0,<5.0.0"
black = ">=25.1.0,<26.0.0"
ruff = ">=0.12.8,<0.13.0"
mypy = ">=1.17.1,<2.0.0"
pytest = ">=8.4.1,<9.0.0"
pytest-cov = "^6"
grpcio-tools = ">=1.74.0,<2.0.0"
pytest-asyncio = ">=1.1.0,<2.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
saop = "saop.cli:main"
